Use a more structured folder and files organization
- Move enpoints to their own folder
- Move services to their own folder
- Move configurations to their own folder
- Move tests to their own folder
