+++ backend/app/main.py
@@ -1,0 +1,24 @@
+from fastapi import FastAPI, HTTPException
+from fastapi.middleware.cors import CORSMiddleware
+from core.config import settings
+from core.exceptions import http_exception_handler, general_exception_handler
+from app.routes.chat import router as chat_router
+
+app = FastAPI()
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=settings.FRONTEND_URL,
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+app.add_exception_handler(HTTPException, http_exception_handler)
+app.add_exception_handler(Exception, general_exception_handler)
+
+app.include_router(chat_router)
+
+@app.get("/")
+def read_root():
+    return {"message": "Hello World"}
