+++ backend/core/config.py
@@ -1,0 +1,13 @@
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+class Settings:
+    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY")
+    FRONTEND_URL: str = "http://localhost:3000"
+
+    class Config:
+        env_file = ".env"
+
+settings = Settings()
