+++ backend/tests/test_chat_endpoint.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from main import app
+from app.main import app
 
 client = TestClient(app)
 
@@ -21,5 +21,4 @@ def test_chat_with_empty_prompt():
 
     response = client.post("/chat/", json=prompt)
 
-    assert response.status_code == 200
-    assert "response" in response.json()
+    assert response.status_code == 422
